+++ src/lib/components/modal/UnsupportedTokensModal.tsx
@@ -49,17 +49,18 @@ const UnsupportedToken = ({ balance }: UnsupportedTokenProps) => {
 
   return (
     <Flex
-      _hover={{
-        "& .info": {
-          display: "flex",
-        },
-      }}
+      className="copier-wrapper"
       borderRadius="8px"
       bg="pebble.800"
       direction="column"
       px={4}
       py={3}
       role="group"
+      _hover={{
+        "& .info": {
+          display: "flex",
+        },
+      }}
     >
       <Flex
         direction="row"
@@ -83,11 +84,11 @@ const UnsupportedToken = ({ balance }: UnsupportedTokenProps) => {
             </Flex>
           </Tooltip>
           <Copier
+            type="unsupported_asset"
             value={balance.id}
             copyLabel="Token ID Copied!"
-            ml="1px"
             display="none"
-            className="info"
+            ml="1px"
           />
         </Flex>
         <Text variant="body3" color="text.dark">