+++ src/lib/pages/past-txs/components/PastTxRow.tsx
@@ -1,5 +1,4 @@
 import { Box, Flex, Grid, Tag, Text, useDisclosure } from "@chakra-ui/react";
-import { useEffect, useState } from "react";
 
 import { RenderActionMessages } from "lib/components/action-msg/ActionMessages";
 import { ExplorerLink } from "lib/components/ExplorerLink";
@@ -20,20 +19,15 @@ export const PastTxRow = ({
   templateColumnsStyle,
 }: PastTxRowProps) => {
   const { isOpen, onToggle } = useDisclosure();
-  const [isAccordion, setIsAccordion] = useState(false);
-  const [showCopyButton, setShowCopyButton] = useState(false);
-  useEffect(() => {
-    if (transaction.messages.length > 1) setIsAccordion(true);
-  }, [transaction.messages]);
+  const isAccordion = transaction.messages.length > 1;
 
   return (
     <Box w="full" minW="min-content">
       <Grid
+        className="copier-wrapper"
         templateColumns={templateColumnsStyle}
         onClick={isAccordion ? onToggle : undefined}
         _hover={{ background: "pebble.900" }}
-        onMouseEnter={() => setShowCopyButton(true)}
-        onMouseLeave={() => setShowCopyButton(false)}
         transition="all .25s ease-in-out"
         cursor={isAccordion ? "pointer" : "default"}
       >
@@ -41,7 +35,7 @@ export const PastTxRow = ({
           <ExplorerLink
             value={transaction.hash.toLocaleUpperCase()}
             type="tx_hash"
-            canCopyWithHover
+            showCopyOnHover
           />
         </TableRow>
         <TableRow>
@@ -53,10 +47,7 @@ export const PastTxRow = ({
         </TableRow>
         <TableRow>
           <Flex gap={1} flexWrap="wrap">
-            <RenderActionMessages
-              transaction={transaction}
-              showCopyButton={showCopyButton}
-            />
+            <RenderActionMessages transaction={transaction} />
             {transaction.isIbc && (
               <Tag borderRadius="full" bg="honeydew.dark" color="pebble.900">
                 IBC